requirements.txt
streamlit
openai
PyPDF2
python-dotenv
google-generativeai